Harden xref/predict parsers; widen Reference.Generation to uint32.
Cap predictor row size at 8 MiB and xref-stream /Size at 1<<24 to
bound memory on adversarial input.  Fix a typo in decodeXRefStream
where the type-field default was keyed on w1 instead of w0, and let
decodeInt reject 8-byte fields whose high bit would overflow int64.
Widen Reference's generation field so out-of-spec values encoded via
xref-stream /W round-trip without silent truncation; classic xref
tables still reject generation > 65535 at write time.

+++ sequential_test.go
@@ -76,7 +76,7 @@ trailer
 
 	type objInfo struct {
 		pos int64
-		gen uint16
+		gen uint32
 	}
 	objs := make(map[uint32]objInfo)
 	for _, sect := range info.Sections {
